Split volume_viewer.py from main.py
